[AIRLFOW-YYY] Move connection commands to separate file
